[WIP] continue modifying the spotify views
