+++ music_controller/spotify/utils.py
@@ -67,6 +67,7 @@ def execute_spotify_api_request(session_id, endpoint, post_=False, put_=False):
         put(BASE_URL + endpoint, headers=headers)
 
     response = get(BASE_URL + endpoint, {}, headers=headers)
+    
     try:
         return response.json()
     except:
